Add extra db population methods for reviews and store reviews
Still have to fix an error in db logic. The food reviews might need to be nested in the store, or the food taken out of the store. I think the former is better since food reviews will never have to be accessed without the food items, and the food items never without the store insofar as I know